+++ models/node.go
@@ -49,6 +49,7 @@ type Node struct {
 	AppEUI        lorawan.EUI64     `db:"app_eui" json:"appEUI"`
 	AppKey        lorawan.AES128Key `db:"app_key" json:"appKey"`
 	UsedDevNonces DevNonceList      `db:"used_dev_nonces" json:"usedDevNonces"`
+	RXDelay       uint8             `db:"rx_delay" json:"rxDelay"`
 }
 
 // ValidateDevNonce returns if the given dev-nonce is valid.